feat: add Ctrl+K command palette, TradingView candlestick charts, SEBI AI compliance guard, automated corporate actions, multi-broker tracking, and Phase 2 ML classifier

+++ frontend/src/components/TransactionDialog.jsx
@@ -25,7 +25,7 @@ const TransactionDialog = ({ holding, transactionType, open, onOpenChange, onSuc
     try {
       const payload = {
         ...formData,
-        quantity: parseInt(formData.quantity),
+        quantity: parseFloat(formData.quantity),
         price: parseFloat(formData.price),
         transaction_type: transactionType,
       };
